commit for me

+++ src/main/java/dznine/projectdb/entity/Buy.java
@@ -1,5 +1,6 @@
 package dznine.projectdb.entity;
 
+
 import javax.persistence.*;
 
 @Entity
@@ -7,7 +8,7 @@ import javax.persistence.*;
 public class Buy {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
     @ManyToOne(fetch = FetchType.LAZY,cascade = CascadeType.ALL)
     @JoinColumn(name = "ordes_id",nullable = false)
     private Orders orders;
@@ -21,11 +22,11 @@ public class Buy {
     @Column(nullable = false)
     private Float total;
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
