+++ src/main/java/dznine/projectdb/entity/Prices.java
@@ -1,7 +1,5 @@
 package dznine.projectdb.entity;
 
-import org.hibernate.annotations.ColumnDefault;
-
 import javax.persistence.*;
 
 @Entity
@@ -9,18 +7,18 @@ import javax.persistence.*;
 public class Prices {
     @Id
     @GeneratedValue
-    private Long id;
+    private Integer id;
     @Column(nullable = false)
     private Float price;
     @ManyToOne(fetch = FetchType.LAZY,cascade = CascadeType.ALL)
     @JoinColumn(name = "products_id", nullable = false)
     private Products products;
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
