+++ src/main/java/dznine/projectdb/repository/BuyRepository.java
@@ -3,5 +3,5 @@ package dznine.projectdb.repository;
 import dznine.projectdb.entity.Buy;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BuyRepository extends JpaRepository<Buy,Long> {
+public interface BuyRepository extends JpaRepository<Buy, Integer> {
 }
