+++ src/main/java/dznine/projectdb/repository/LogsActionRepository.java
@@ -3,5 +3,5 @@ package dznine.projectdb.repository;
 import dznine.projectdb.entity.LogsActions;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LogsActionRepository extends JpaRepository<LogsActions,Long> {
+public interface LogsActionRepository extends JpaRepository<LogsActions, Integer> {
 }
