+++ src/main/java/dznine/projectdb/repository/OrdersActionsRepository.java
@@ -3,5 +3,5 @@ package dznine.projectdb.repository;
 import dznine.projectdb.entity.OrdersActions;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrdersActionsRepository extends JpaRepository<OrdersActions,Long> {
+public interface OrdersActionsRepository extends JpaRepository<OrdersActions, Integer> {
 }
