+++ src/main/java/dznine/projectdb/repository/SalesRepository.java
@@ -3,5 +3,5 @@ package dznine.projectdb.repository;
 import dznine.projectdb.entity.Sales;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SalesRepository extends JpaRepository<Sales,Long> {
+public interface SalesRepository extends JpaRepository<Sales, Integer> {
 }
