+++ src/main/java/dznine/projectdb/repository/UnitsRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UnitsRepository extends JpaRepository<Units,Long> {
+public interface UnitsRepository extends JpaRepository<Units, Integer> {
 }
